errors

+++ controllers/card.js
@@ -1,59 +1,63 @@
 const mongoose = require('mongoose');
 const Card = require('../models/card');
-const errors = require('../utils/errorCard');
 
 const { DocumentNotFoundError } = mongoose.Error;
 
-const getCard = (req, res) => {
+const getCard = (req, res, next) => {
   Card.find({})
     .populate(['owner', 'likes'])
     .then((cards) => res.status(200).send({ data: cards }))
-    .catch((e) => errors(e, res));
+    .catch((err) => next(err));
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
   Card.create({ name, link, owner })
     .then((card) => res.status(201).send({ data: card }))
-    .catch((e) => {
-      const errMessage = 'Переданы некорректные данные при создании карточки.';
-      errors(e, res, errMessage);
-    });
+    .catch((err) => next(err));
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(new DocumentNotFoundError())
-    .then((card) => res.send({ data: card }))
-    .catch((e) => errors(e, res));
+const deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(new DocumentNotFoundError('Карточка с указанным _id не найдена'))
+    .then((card) => {
+      if (card.owner.toString() !== req.user._id) {
+        next(new DocumentNotFoundError('Доступ запрещен'));
+      }
+      Card.findByIdAndRemove(req.params.cardId);
+      res.send({ data: card });
+    })
+    .catch((err) => {
+      next(err);
+    });
 };
-const likeDeleteCard = (req, res, keyMethod) => {
+const likeDeleteCard = (req, res, next, keyMethod) => {
   Card
     .findByIdAndUpdate(
       req.params.cardId,
       keyMethod,
       { new: true },
     )
-    .orFail(new DocumentNotFoundError())
+    .orFail(new DocumentNotFoundError('Карточка с указанным _id не найдена'))
     .then((card) => card.populate(['owner', 'likes']))
     .then((card) => {
       res.send({ data: card });
     })
-    .catch((e) => errors(e, res));
+    .catch((err) => next(err));
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   // добавить _id в массив, если его там нет
   const keyMethod = { $addToSet: { likes: req.user._id } };
-  likeDeleteCard(req, res, keyMethod);
+  likeDeleteCard(req, res, next, keyMethod);
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   // убрать _id из массива
   const keyMethod = { $pull: { likes: req.user._id } };
-  likeDeleteCard(req, res, keyMethod);
+  likeDeleteCard(req, res, next, keyMethod);
 };
 
 module.exports = {
